Add audio focus request feature.
When the call goes in CALL_STATE_CONNECTED the audio focus is requested. When the call is ended (CALL_STATE_ENDED) the focus is released.
If a media player is playing, according to the implementation the music is stopped. Also when the audio focus is released, the media player does not resume the music.