remove # before build version
Every time when people refer to a riot-build they use the string
as provided in the app. This leads to a trigger of the github bot
which thinks that the version string refers to a issue/ pull request

To avoid this we replace the hashtag with a b so we now get a version
like G-b1234. This is currently not used anywhere as trigger (AFAIK)
so the room should not be polluted with content of overactive bots